requirements.txt
# RePRO-Agent: Complete Requirements
# Scientific Reproducibility and Hypothesis Generation Pipeline

# Core dependencies
argparse
json
logging
pathlib
typing
datetime
tempfile
shutil
subprocess
os
sys
uuid
collections

# PDF and document processing
PyMuPDF>=1.23.0
unstructured[pdf]>=0.10.0

# LLM APIs
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Web requests and utilities
requests>=2.31.0
python-dotenv>=1.0.0

# Git operations
GitPython>=3.1.0

# Docker integration
docker>=6.0.0

# YAML processing
PyYAML>=6.0

# Python dependency analysis
pipreqs>=0.4.0

# RDF and knowledge graphs
rdflib>=6.0.0
SPARQLWrapper>=2.0.0

# Date utilities
python-dateutil>=2.8.0

# Network analysis
networkx>=3.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
